// src/app/core/models/enums/order-status.ts
export enum OrderStatus {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  COMPLETED = 'COMPLETED',
  CONFIRMED = 'CONFIRMED',
  PAID = 'PAID',
  INDELIVERY = 'INDELIVERY',
  REJECTED = 'REJECTED',
  VALIDATED = 'VALIDATED',
  CANCELLED = 'CANCELLED',
  FAILED = 'FAILED',
  AWAITING_CASH = 'AWAITING_CASH'
}